Refactor metrics as external dependency to projects
