--- addresspoints/src/main/scala/grasshopper/addresspoints/metrics/Instrumented.scala
@@ -1,7 +1,0 @@
-package grasshopper.addresspoints.metrics
-
-import nl.grons.metrics.scala.InstrumentedBuilder
-
-trait Instrumented extends InstrumentedBuilder {
-  val metricRegistry = Registry.metricsRegistry
-}
