+++ census/src/main/scala/grasshopper/census/CensusGeocodeService.scala
@@ -5,7 +5,7 @@ import akka.event.Logging
 import akka.http.scaladsl.Http
 import akka.stream.ActorMaterializer
 import com.typesafe.config.ConfigFactory
-import grasshopper.census.metrics.JvmMetrics
+import grasshopper.metrics.JvmMetrics
 import org.elasticsearch.client.transport.TransportClient
 import org.elasticsearch.common.transport.InetSocketTransportAddress
 import grasshopper.census.api.Service
