--- census/src/main/scala/grasshopper/census/metrics/Instrumented.scala
@@ -1,7 +1,0 @@
-package grasshopper.census.metrics
-
-import nl.grons.metrics.scala.InstrumentedBuilder
-
-trait Instrumented extends InstrumentedBuilder {
-  val metricRegistry = Registry.metricsRegistry
-}
