--- census/src/main/scala/grasshopper/census/metrics/Registry.scala
@@ -1,7 +1,0 @@
-package grasshopper.census.metrics
-
-import com.codahale.metrics.MetricRegistry
-
-object Registry {
-  val metricsRegistry = new MetricRegistry
-}
