--- geocoder/src/main/scala/grasshopper/geocoder/metrics/Registry.scala
@@ -1,7 +1,0 @@
-package grasshopper.geocoder.metrics
-
-import com.codahale.metrics.MetricRegistry
-
-object Registry {
-  val metricsRegistry = new MetricRegistry
-}
