+++ metrics/src/main/resources/application.conf
@@ -1,0 +1,12 @@
+grasshopper {
+  geocoder {
+     monitoring {
+      isMonitored = false
+      frequency = 10
+      influxdb {
+        host = "localhost"
+        port = "8086"
+      }
+    }
+  }
+}
