+++ project/GrasshopperBuild.scala
@@ -43,7 +43,7 @@ object GrasshopperBuild extends Build {
 
   val scaleDeps = Seq(scaleGeoJson)
 
-  val metricsDeps = Seq(metrics, metricsJvm, influxDbReporter)
+  val metricsDeps = Seq(config, metricsScala, metricsJvm, influxDbReporter)
 
   val geocodeDeps = akkaHttpDeps ++ esDeps ++ scaleDeps ++ metricsDeps
 
@@ -65,6 +65,18 @@ object GrasshopperBuild extends Build {
       )
     )
 
+  lazy val metrics = (project in file("metrics"))
+    .configs(IntegrationTest)
+    .settings(buildSettings: _*)
+    .settings(
+      Revolver.settings ++
+      Seq(
+        assemblyJarName in assembly := {s"grasshopper-${name.value}.jar"},
+        libraryDependencies ++= metricsDeps,
+        resolvers ++= repos
+      )
+    )
+
   lazy val addresspoints = (project in file("addresspoints"))
     .configs( IntegrationTest )
     .settings(buildSettings: _*)
@@ -75,7 +87,7 @@ object GrasshopperBuild extends Build {
         libraryDependencies ++= geocodeDeps,
         resolvers ++= repos
       )
-    ).dependsOn(elasticsearch)
+    ).dependsOn(elasticsearch, metrics)
 
   lazy val census = (project in file("census"))
     .configs( IntegrationTest )
@@ -87,7 +99,7 @@ object GrasshopperBuild extends Build {
         libraryDependencies ++= geocodeDeps,
         resolvers ++= repos
       )
-    ).dependsOn(elasticsearch)
+    ).dependsOn(elasticsearch, metrics)
 
   lazy val client = (project in file("client"))
     .configs( IntegrationTest )
@@ -116,9 +128,7 @@ object GrasshopperBuild extends Build {
         libraryDependencies ++= akkaHttpDeps ++ scaleDeps ++ asyncDeps ++ metricsDeps,
         resolvers ++= repos
       )
-    ).dependsOn(client)
-
-
+    ).dependsOn(client, metrics)
 
 
 }
